+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 - Fix issue loading instances when there were none
 - Fix downloads redownloading without a hash, but size matches
 - Turn off mutex in exe
+- Fix organising filesystem in linux